Issue-93: Merged SingleResponseFactory and MultipleResponseFactory together.
In effort to reduce the complexity of interacting with the Stumps API, the
factories SingleResponseFactory and MultipleResponseFactory are now merged
together into a single StumpResponseFactory based on the IStumpResponseFactory
interface. This is accomplished since the default behavior of
MultipleResponseFactory was to infinitely loop across responses, a single
response within the array of responses would produce the same result.

The response factory itself therefore could have been removed, but its
existence opens the possibility for more advanced responses in the future
wherein the response factory could dynamically generate a response using
a rendering framework such as Mustache or Razor.