Issue-94: Moved drop connection and time delay features to the HTTP response.
Moving the drop connection and time delay features to the HTTP response allows
the a multi-response stump have a different behavior for each individual
request/response rather than globally for the entire stump.  This can be useful
when trying to simulate chaotic network conditions for a single request.

Previously this was only possible using multiple stumps with input
specifically required by the client to trigger the behavior.

** BREAKING CHANGE **

The API has now changed, as has the JSON serialization.